[Silver II] Title: 랜선 자르기, Time: 4 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.c
@@ -5,7 +5,7 @@ int main() {
     
     scanf("%d %d", &n, &k);
     
-    long wires[n];
+    int wires[n];
     long start, end = 1;
     
     for (int i = 0; i < n; i++) {
